requirements.txt
streamlit==1.30.0
azure-ai-contentsafety==1.0.0
openai==1.6.1
python-dotenv==1.0.0
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
pandas==2.1.1
PyPDF2==3.0.1
python-docx==1.0.1
